+++ Sources/SwiftMCP/Transport/Routing/OAuthRoutes.swift
@@ -363,20 +363,3 @@ extension HTTPSSETransport {
 		return "\(scheme)://\(host)"
 	}
 }
-
-
-// MARK: - URLSession Delegate
-
-/// A URLSession delegate that prevents automatic redirect following,
-/// allowing the proxy to return redirect responses to the client.
-private final class NoRedirectDelegate: NSObject, URLSessionTaskDelegate {
-	func urlSession(
-		_ session: URLSession,
-		task: URLSessionTask,
-		willPerformHTTPRedirection response: HTTPURLResponse,
-		newRequest request: URLRequest,
-		completionHandler: @escaping (URLRequest?) -> Void
-	) {
-		completionHandler(nil) // Don't follow redirects
-	}
-}
