+++ Sources/SwiftMCP/Transport/Routing/RouteMatch.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// Result of a successful route match.
+struct RouteMatch: Sendable {
+	let route: HTTPRoute
+	let pathParams: [String: String]
+}
